DP-1169: Move footer to new folder, app
